+++ Readiculous.WebApp/Controllers/UserMasterController.cs
@@ -31,8 +31,16 @@ namespace Readiculous.WebApp.Controllers
             _userService = userService;
         }
 
+        private bool IsReviewer()
+        {
+            var user = _userService.GetUserEditById(this.UserId);
+            return user.Role == Readiculous.Resources.Constants.Enums.RoleType.Reviewer;
+        }
+
         public IActionResult UserMasterScreen(string searchString, RoleType? roleType, UserSortType searchType, int page = 1, int pageSize = 10)
         {
+            if (IsReviewer())
+                return View("~/Views/Shared/Forbidden.cshtml");
             ViewData["CurrentFilter"] = searchString;
             ViewData["CurrentRoleType"] = roleType.HasValue ? roleType.Value : string.Empty;
             ViewData["CurrentUserSearchType"] = searchType.ToString();
@@ -57,12 +65,16 @@ namespace Readiculous.WebApp.Controllers
         [HttpGet]
         public IActionResult UserAddModal()
         {
+            if (IsReviewer())
+                return View("~/Views/Shared/Forbidden.cshtml");
             return PartialView(new UserViewModel());
         }
 
         [HttpPost]
         public async Task<IActionResult> Create(UserViewModel model)
         {
+            if (IsReviewer())
+                return View("~/Views/Shared/Forbidden.cshtml");
             if (ModelState.IsValid)
             {
                 try
@@ -81,6 +93,8 @@ namespace Readiculous.WebApp.Controllers
         [HttpGet]
         public IActionResult UserEditModal(string userId)
         {
+            if (IsReviewer())
+                return View("~/Views/Shared/Forbidden.cshtml");
             try
             {
                 var user = _userService.GetUserEditById(userId);
@@ -95,6 +109,8 @@ namespace Readiculous.WebApp.Controllers
         [HttpPost]
         public async Task<IActionResult> Edit(UserViewModel model)
         {
+            if (IsReviewer())
+                return View("~/Views/Shared/Forbidden.cshtml");
             if (ModelState.IsValid)
             {
                 try
@@ -113,6 +129,8 @@ namespace Readiculous.WebApp.Controllers
         [HttpGet]
         public IActionResult UserViewModal(string userId)
         {
+            if (IsReviewer())
+                return View("~/Views/Shared/Forbidden.cshtml");
             try
             {
                 var user = _userService.GetUserDetailsById(userId);
@@ -127,6 +145,8 @@ namespace Readiculous.WebApp.Controllers
         [HttpPost]
         public async Task<IActionResult> Delete(string userId)
         {
+            if (IsReviewer())
+                return View("~/Views/Shared/Forbidden.cshtml");
             try
             {
                 await _userService.DeleteUserAsync(userId, this.UserId);
